# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/export/fightcard/pdf": {
      "post": {
        "operationId": "ExternalServicesController_getFightCardPdf",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportWithQrCodeDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/fightcard/png": {
      "post": {
        "operationId": "ExternalServicesController_getFightCardPng",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportWithQrCodeDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/selector/battlecard": {
      "post": {
        "operationId": "ExternalServicesController_getBattlecard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorExportDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/fightcard/csv": {
      "post": {
        "operationId": "ExternalServicesController_getFightCardCsv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleTournamentDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/fightcard/xlsx": {
      "post": {
        "operationId": "ExternalServicesController_getFightCardXlsx",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleTournamentDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/fightcard/html": {
      "get": {
        "operationId": "ExternalServicesController_getTemplateHtml",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {}
        ],
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/selector/pdf": {
      "post": {
        "operationId": "ExternalServicesController_getSelectorPdf",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorExportDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/selector/png": {
      "post": {
        "operationId": "ExternalServicesController_getSelectorPng",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorExportDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/selector/csv": {
      "post": {
        "operationId": "ExternalServicesController_getSelectorCsv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorExportDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/export/selector/xlsx": {
      "post": {
        "operationId": "ExternalServicesController_getSelectorXlsx",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorExportDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "ExternalServices"
        ]
      }
    },
    "/api/import": {
      "post": {
        "operationId": "ImportController_importBoxers",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportBoxersDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBoxersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Import"
        ]
      }
    },
    "/api/import/previewfromcsvText": {
      "post": {
        "operationId": "ImportController_previewBoxersFromCsvText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewBoxersCsvDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewBoxersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Import"
        ]
      }
    },
    "/api/import/previewFromCsvFile": {
      "post": {
        "operationId": "ImportController_previewBoxersFromCsvFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewBoxersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Import"
        ]
      }
    },
    "/api/import/previewfromffboxeFile": {
      "post": {
        "operationId": "ImportController_previewBoxersFromFfboxeFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewBoxersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Import"
        ]
      }
    },
    "/api/import/previewFromApi": {
      "post": {
        "operationId": "ImportController_previewFromApi",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewBoxersApiDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewBoxersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Import"
        ]
      }
    },
...
  },
  "info": {
    "title": "Battlecard",
    "description": "The Fightmaker App",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Gender": {
        "type": "string",
        "enum": [
          "female",
          "male"
        ],
        "description": "Gender"
      },
...
      "ExportWithQrCodeDto": {
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string"
          },
          "displayQrCode": {
            "type": "boolean"
          }
        },
        "required": [
          "tournamentId",
          "displayQrCode"
        ]
      },
      "SelectorExportDto": {
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string"
          },
          "boxerIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "tournamentId",
          "boxerIds"
        ]
      },
      "SimpleTournamentDto": {
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string"
          }
        },
        "required": [
          "tournamentId"
        ]
      },
      "ImportBoxerDto": {
        "type": "object",
        "properties": {
          "gender": {
            "description": "Boxer gender",
            "allOf": [
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "lastName": {
            "type": "string",
            "description": "Boxer name"
          },
          "firstName": {
            "type": "string",
            "description": "Boxer first name"
          },
          "birthDate": {
            "type": "string",
            "description": "Boxer birth date (YYYY-MM-DD)"
          },
          "club": {
            "type": "string",
            "description": "Boxer club"
          },
          "weight": {
            "type": "number",
            "description": "Boxer weight"
          },
          "license": {
            "type": "string",
            "description": "Boxer license"
          },
          "fightRecord": {
            "type": "number",
            "description": "Boxer fight record"
          }
        },
        "required": [
          "lastName",
          "firstName",
          "birthDate",
          "club",
          "weight",
          "license",
          "fightRecord"
        ]
      },
      "ImportBoxersDto": {
        "type": "object",
        "properties": {
          "dry": {
            "type": "boolean",
            "description": "Whether this is a dry run"
          },
          "boxers": {
            "description": "Array of boxers to import",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportBoxerDto"
            }
          },
          "tournamentId": {
            "type": "string",
            "description": "Tournament ID"
          }
        },
        "required": [
          "dry",
          "boxers",
          "tournamentId"
        ]
      },
      "BoxerImportErrorDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "row": {
            "type": "number",
            "description": "Row number where error occurred"
          },
          "field": {
            "type": "string",
            "description": "Field name that caused the error"
          }
        },
        "required": [
          "message",
          "row",
          "field"
        ]
      },
      "ImportBoxersResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the import was successful"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "errors": {
            "description": "Array of import errors",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoxerImportErrorDto"
            }
          }
        },
        "required": [
          "success",
          "message",
          "errors"
        ]
      },
      "PreviewBoxersCsvDto": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "string",
            "description": "CSV payload"
          }
        },
        "required": [
          "payload"
        ]
      },
      "ImportBoxerResponseDto": {
        "type": "object",
        "properties": {
          "gender": {
            "description": "Boxer gender",
            "allOf": [
              {
                "$ref": "#/components/schemas/Gender"
              }
            ]
          },
          "lastName": {
            "type": "string",
            "description": "Boxer name"
          },
          "firstName": {
            "type": "string",
            "description": "Boxer first name"
          },
          "birthDate": {
            "type": "string",
            "description": "Boxer birth date (YYYY-MM-DD)"
          },
          "club": {
            "type": "string",
            "description": "Boxer club"
          },
          "weight": {
            "type": "number",
            "description": "Boxer weight"
          },
          "license": {
            "type": "string",
            "description": "Boxer license"
          },
          "fightRecord": {
            "type": "number",
            "description": "Boxer fight record"
          }
        },
        "required": [
          "lastName",
          "firstName",
          "birthDate",
          "club",
          "weight",
          "license",
          "fightRecord"
        ]
      },
      "PreviewBoxersResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the preview was successful"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "boxers": {
            "description": "Array of preview boxers",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportBoxerResponseDto"
            }
          }
        },
        "required": [
          "success",
          "message",
          "boxers"
        ]
      },
      "FileUploadDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "file"
        ]
      },
      "PreviewBoxersApiDto": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "string",
            "description": "Payload containing boxer IDs"
          }
        },
        "required": [
          "payload"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}